+++ Runtime/Scripts/Base/Singleton.cs
@@ -6,7 +6,7 @@
  * Date Created: April 28, 2022
  * 
  * Additional Comments:
- *      File Line Length: 120
+ *      File Line Length: 140
  */
 
 using UnityEngine;